requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
importlib-metadata==4.5.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1

PyYAML~=5.4.1
numpy~=1.20.3
selenium~=3.141.0
Faker~=8.8.2
Appium-Python-Client~=1.2.0